+++ lib/ts/recipe/emailverification/recipeImplementation.ts
@@ -44,7 +44,7 @@ export default class RecipeImplementation implements RecipeInterface {
         const response: SendVerifyEmailAPIResponse = await this.querier.post(
             "/user/email/verify/token",
             {},
-            (context) => {
+            async (context) => {
                 return input.config.preAPIHook({
                     ...context,
                     action: "SEND_VERIFY_EMAIL",
@@ -66,7 +66,7 @@ export default class RecipeImplementation implements RecipeInterface {
             "/user/email/verify",
             {},
             undefined,
-            (context) => {
+            async (context) => {
                 return input.config.preAPIHook({
                     ...context,
                     action: "IS_EMAIL_VERIFIED",