+++ lib/ts/recipe/recipeModule/types.ts
@@ -2,7 +2,7 @@ import { NormalisedAppInfo } from "../../types";
 
 export type UserInput<GetRedirectionURLContextType, PreAPIHookContextType, OnHandleEventContextType> = {
     getRedirectionURL?: (context: GetRedirectionURLContextType) => Promise<string | undefined>;
-    preAPIHook?: (context: PreAPIHookContextType) => Promise<RequestInit | { url?: string; requestInit: RequestInit }>;
+    preAPIHook?: (context: PreAPIHookContextType) => Promise<{ url: string; requestInit: RequestInit }>;
     onHandleEvent?: (context: OnHandleEventContextType) => void;
     useShadowDom?: boolean;
     palette?: Record<string, string>;
@@ -17,7 +17,7 @@ export type NormalisedConfig<GetRedirectionURLContextType, PreAPIHookContextType
     recipeId: string;
     appInfo: NormalisedAppInfo;
     getRedirectionURL: (context: GetRedirectionURLContextType) => Promise<string | undefined>;
-    preAPIHook: (context: PreAPIHookContextType) => Promise<RequestInit | { url?: string; requestInit: RequestInit }>;
+    preAPIHook: (context: PreAPIHookContextType) => Promise<{ url: string; requestInit: RequestInit }>;
     onHandleEvent: (context: OnHandleEventContextType) => void;
     useShadowDom: boolean;
     palette: Record<string, string>;