+++ lib/ts/recipe/session/types.ts
@@ -13,39 +13,29 @@
  * under the License.
  */
 
-import { Config as RecipeModuleConfig } from "../recipeModule/types";
-import RecipeImplementation from "./recipeImplementation";
+import { RecipeImplementation, RecipeInterface } from "supertokens-website";
 
-/*
- * Session User InputsConfig Types.
- */
-
-export type UserInput = {
+export type InputType = {
+    apiDomain?: string;
+    apiBasePath?: string;
     sessionScope?: string;
-    refreshAPICustomHeaders?: any;
-    signoutAPICustomHeaders?: any;
     sessionExpiredStatusCode?: number;
     autoAddCredentials?: boolean;
     isInIframe?: boolean;
+    cookieDomain?: string;
+    preAPIHook?: (context: {
+        action: "SIGN_OUT" | "REFRESH_SESSION";
+        requestInit: RequestInit;
+        url: string;
+    }) => Promise<{ url: string; requestInit: RequestInit }>;
+    onHandleEvent?: (context: { action: "SIGN_OUT" | "REFRESH_SESSION" | "UNAUTHORISED" }) => void;
     override?: {
         functions?: (originalImplementation: RecipeImplementation) => RecipeInterface;
     };
 };
-
-export type Config = UserInput & RecipeModuleConfig<unknown, unknown, unknown>;
 
 export type SessionContextType = {
     doesSessionExist: boolean;
     userId: string;
     jwtPayload: any;
 };
-
-export interface RecipeInterface {
-    getUserId(): Promise<string>;
-
-    getJWTPayloadSecurely(): Promise<any>;
-
-    doesSessionExist(): Promise<boolean>;
-
-    signOut(): Promise<void>;
-}